+++ client-app/app.config.ts
@@ -10,6 +10,7 @@ export default {
     assetBundlePatterns: ["**/*"],
     orientation: "portrait",
     icon: "./assets/icon.png",
+    scheme: "la-historical-markers",
     ios: {
       icon: "./assets/icon.png",
       bundleIdentifier: "com.austinwebre.lahistoricalmarkers",
