+++ client-app/package.json
@@ -9,11 +9,15 @@
   },
   "dependencies": {
     "@react-native-async-storage/async-storage": "^1.13.0",
+    "@react-native-community/masked-view": "0.1.10",
+    "@react-navigation/native": "^5.9.4",
+    "@react-navigation/stack": "^5.14.4",
     "babel-plugin-module-resolver": "^4.1.0",
     "expo": "^41.0.0",
     "expo-constants": "~10.1.3",
     "expo-image-manipulator": "~9.1.0",
     "expo-image-picker": "~10.1.3",
+    "expo-linking": "~2.2.3",
     "expo-location": "~12.0.4",
     "expo-splash-screen": "~0.10.2",
     "expo-updates": "~0.5.4",
@@ -28,6 +32,7 @@
     "react-native-maps": "0.27.1",
     "react-native-picker-select": "^8.0.4",
     "react-native-reanimated": "~2.1.0",
+    "react-native-safe-area-context": "3.2.0",
     "react-native-screens": "~3.0.0",
     "react-native-svg": "^12.1.1",
     "react-native-web": "~0.13.12",
@@ -39,6 +44,7 @@
     "@types/react-dom": "~16.9.8",
     "@types/react-native": "~0.63.2",
     "babel-preset-expo": "8.3.0",
+    "expo-cli": "^4.4.3",
     "jest-expo": "^41.0.0",
     "typescript": "~4.0.0"
   },
