+++ client-app/src/screens/AdminScreen.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import { Text } from "react-native";
+
+export default function AdminScreen() {
+  return <Text>Admin!</Text>;
+}
