+++ client-app/src/screens/index.tsx
@@ -1,0 +1,3 @@
+import MarkersScreen from "./MarkersScreen";
+
+export { MarkersScreen };
